switch regex_search to sregex_iterator
Matching using regex_search requires to loop and search through the
suffix of each match. If a match was found at the beginning of a word,
the word would be spliced and, therefore, word boundaries such as '\b'
at the start of a regex would not be respected. Using sregex_iterator
fixes this issue.